rust: Simplify example loaders
