+++ rust/tests/lib.rs
@@ -4,12 +4,13 @@ use zser::decoder::Decoder;
 use zser::parser::Parser;
 use zser::varint;
 
-mod message_examples;
-mod varint_examples;
+mod test_vectors;
+
+use test_vectors::{MessageExample, VarintExample, decode_value};
 
 #[test]
 fn message_examples() {
-    let examples = message_examples::load();
+    let examples = MessageExample::load_all();
 
     for example in examples {
         let mut parser = Parser::new(Decoder::new());
@@ -22,7 +23,7 @@ fn message_examples() {
             }
 
             let value = parser.finish().expect("finished");
-            let expected = message_examples::decode_value(&example.decoded.expect("decoded"));
+            let expected = decode_value(&example.decoded.expect("decoded"));
             assert_eq!(value, expected);
         } else if parser.parse(&example.encoded).is_ok() {
             panic!("{}: expected error but example parsed successfully",
@@ -33,7 +34,7 @@ fn message_examples() {
 
 #[test]
 fn varint_encode() {
-    let examples = varint_examples::load();
+    let examples = VarintExample::load_all();
     let mut output = [0u8; 9];
 
     for example in examples {
@@ -48,7 +49,7 @@ fn varint_encode() {
 
 #[test]
 fn varint_decode() {
-    let examples = varint_examples::load();
+    let examples = VarintExample::load_all();
 
     for example in examples {
         let mut encoded_ref = &example.encoded[..];
